requirements.txt
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
typing-extensions==4.8.0
urllib3==2.0.7
werkzeug==3.0.1
zipp==3.17.0
